refactor: extract useSwipeNavigation — swipe/greeting navigation logic
