+++ src/views/ChatView.vue
@@ -97,6 +97,7 @@ import { useContextBreakdown } from '@/composables/chat/useContextBreakdown.js';
 import { useMessageSelection } from '@/composables/chat/useMessageSelection.js';
 import { useChatSearch } from '@/composables/chat/useChatSearch.js';
 import { useMemorySheetUI } from '@/composables/chat/useMemorySheetUI.js';
+import { useSwipeNavigation } from '@/composables/chat/useSwipeNavigation.js';
 import { normalizeImgGenHtmlForEditing, prepareEditText, restoreEditText } from '@/core/utils/messageEditHelpers.js';
 
 function genMsgId() {
@@ -1640,97 +1641,15 @@ async function branchSession(msgIndex) {
 
 // --- Message Actions ---
 
-function changeSwipe(msgIndex, dir, fromSwipe = false) {
-    if (isGenerating.value) return;
-    const msg = currentMessages.value[msgIndex];
-    
-    // If the current message is an error and other swipes exist, remove the error entry on swipe
-    if (msg.isError && msg.swipes && msg.swipes.length > 1) {
-        const errorSwipeId = msg.swipeId || 0;
-        
-        msg.swipes.splice(errorSwipeId, 1);
-        if (msg.swipesMeta) msg.swipesMeta.splice(errorSwipeId, 1);
-        
-        let newIndex = errorSwipeId;
-        if (dir < 0) newIndex = errorSwipeId - 1;
-        
-        if (newIndex >= msg.swipes.length) newIndex = msg.swipes.length - 1;
-        if (newIndex < 0) newIndex = 0;
-        
-        msg.swipeId = newIndex;
-        msg.text = msg.swipes[newIndex];
-        msg.isError = false;
-        msg.swipeDirection = fromSwipe ? (dir > 0 ? 'slide-next' : 'slide-prev') : 'fade';
-        
-        let newReasoning = null;
-        let newGenTime = null;
-        let newTokens = null;
-        if (msg.swipesMeta && msg.swipesMeta[newIndex]) {
-            newReasoning = msg.swipesMeta[newIndex].reasoning;
-            newGenTime = msg.swipesMeta[newIndex].genTime;
-            newTokens = msg.swipesMeta[newIndex].tokens;
-        }
-        msg.reasoning = newReasoning;
-        msg.genTime = newGenTime;
-        msg.tokens = newTokens;
-        
-        updateSessionMessage(activeChatChar, msgIndex, msg);
-        return;
-    }
-
-    if (!msg.swipes || msg.swipes.length <= 1) return;
-    
-    let newIndex = (msg.swipeId || 0) + dir;
-
-    const isLastMsg = msgIndex === currentMessages.value.length - 1;
-    if (dir > 0 && newIndex >= msg.swipes.length && isLastMsg) {
-        regenerateMessage(msgIndex, 'new_variant');
-        return;
-    }
-    
-    if (newIndex < 0 || newIndex >= msg.swipes.length) return;
-    
-    msg.swipeDirection = fromSwipe ? (dir > 0 ? 'slide-next' : 'slide-prev') : 'fade';
-    msg.swipeId = newIndex;
-    msg.text = msg.swipes[newIndex];
-    msg.isError = false;
-
-    let newReasoning = null;
-    let newGenTime = null;
-    let newTokens = null;
-    if (msg.swipesMeta && msg.swipesMeta[newIndex]) {
-        newReasoning = msg.swipesMeta[newIndex].reasoning;
-        newGenTime = msg.swipesMeta[newIndex].genTime;
-        newTokens = msg.swipesMeta[newIndex].tokens;
-    }
-    msg.reasoning = newReasoning;
-    msg.genTime = newGenTime;
-    msg.tokens = newTokens;
-
-    updateSessionMessage(activeChatChar, msgIndex, msg);
-}
-
-function changeGreeting(msgIndex, dir, fromSwipe = false) {
-    if (isGenerating.value) return;
-    const msg = currentMessages.value[msgIndex];
-    const persona = activePersona.value;
-    const greetings = getAllGreetings(activeChatChar, persona);
-    if (greetings.length <= 1) return;
-    
-    let newIndex = (msg.greetingIndex || 0) + dir;
-    if (newIndex >= greetings.length) newIndex = 0;
-    if (newIndex < 0) newIndex = greetings.length - 1;
-    
-    msg.swipeDirection = fromSwipe ? (dir > 0 ? 'slide-next' : 'slide-prev') : 'fade';
-    msg.greetingIndex = newIndex;
-    msg.text = greetings[newIndex];
-    msg.tokens = estimateTokens(greetings[newIndex]);
-    msg.swipes = [msg.text]; // Reset swipes for greeting
-    msg.swipeId = 0;
-    msg.reasoning = null;
-    msg.isError = false;
-    updateSessionMessage(activeChatChar, msgIndex, msg);
-}
+const {
+    changeSwipe,
+    changeGreeting
+} = useSwipeNavigation({
+    currentMessages,
+    isGenerating,
+    getActiveChatChar: () => activeChatChar,
+    regenerateMessage: (msgIndex, mode, guidanceText) => regenerateMessage(msgIndex, mode, guidanceText)
+});
 
 function getReasoningTags() {
     let { start, end } = getApiReasoningTags();
